Wasm

+++ crates/app_core/src/view/projects_view.rs
@@ -5,17 +5,17 @@ use super::project_with_thumbnail_view::ProjectWithThumbnailView;
 #[derive(Serialize, Deserialize, Default, Debug, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct ProjectsView {
-    pub page: i32,
-    pub size: i32,
-    pub total_pages: i32,
+    pub page: i64,
+    pub size: i64,
+    pub total_pages: i64,
     pub projects: Vec<ProjectWithThumbnailView>,
 }
 
 impl ProjectsView {
     pub fn new(
-        page: i32,
-        size: i32,
-        total_pages: i32,
+        page: i64,
+        size: i64,
+        total_pages: i64,
         projects: Vec<ProjectWithThumbnailView>,
     ) -> Self {
         Self {
